move glean to guest layout instead of home page, and use UUID derived from FxA UID
