+++ .env-dist
@@ -151,3 +151,6 @@ NEXT_PUBLIC_MONITOR_SUPPORT_URL=https://support.mozilla.org/kb/firefox-monitor
 GCP_PUBSUB_PROJECT_ID=
 GCP_PUBSUB_TOPIC_NAME=
 GCP_PUBSUB_SUBSCRIPTION_NAME=
+
+// Randomly-generated UUIDv5 namespace, until/unless we are approved to use Fxa UID for Nimbus User ID.
+NIMBUS_UUID_NAMESPACE=
