+++ src/app/(nextjs_migration)/(guest)/page.tsx
@@ -13,43 +13,9 @@ import LockImage from "../../../client/images/landing-lock@2x.webp";
 import MailImage from "../../../client/images/landing-mail@2x.webp";
 import NaturePhoneImage from "../../../client/images/landing-nature-phone@2x.webp";
 import { getNonce } from "../functions/server/getNonce";
-import { PageLoadEvent } from "../components/client/PageLoadEvent";
-import { getServerSession } from "next-auth";
-import { authOptions } from "../../api/utils/auth";
-import { isFlagEnabled } from "../../functions/server/featureFlags";
-import { cookies } from "next/headers";
-import { getRandomValues } from "crypto";
 
-export default async function Home() {
+export default function Home() {
   const l10n = getL10n();
-
-  // If the user is logged in, use subscriber ID as Nimbus user ID.
-  const session = await getServerSession(authOptions);
-  let userId = session?.user?.subscriber?.id?.toString();
-
-  // if the user is not logged in, use a cookie with a randomly-generated Nimbus user ID.
-  if (!userId) {
-    const cookie = cookies().get("userId");
-    if (cookie) {
-      userId = cookie.value;
-    } else {
-      const array = new Uint32Array(1);
-      getRandomValues(array);
-      // TODO Cookies can only be set in server action or route handler
-      // @see https://nextjs.org/docs/app/api-reference/functions/cookies#cookiessetname-value-options
-      // cookies().set("userId", uuid);
-      userId = array.toString();
-    }
-  }
-
-  // @see https://github.com/mozilla/experimenter/tree/main/cirrus
-  const serverUrl = process.env.SERVER_URL ?? "http://localhost:6060";
-
-  //@ts-ignore TODO this tells us which features to enable.
-  const features = await fetch(`${serverUrl}/v1/features/`, {
-    method: "POST",
-    body: JSON.stringify({ client_id: userId }),
-  });
 
   return (
     <div data-partial="landing">
@@ -68,7 +34,6 @@ export default async function Home() {
         src="/nextjs_migration/client/js/landing.js"
         nonce={getNonce()}
       />
-      <PageLoadEvent userId={userId} />
       <section className="hero">
         <div>
           <h1>{l10n.getString("exposure-landing-hero-heading")}</h1>
